requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-multipart
boto3
python-jose[cryptography]
# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
# Async HTTP (Google OAuth token exchange)
httpx>=0.27
# Vector DB (RAG queries only — ingestion lives in celery_worker)
pymilvus
sentence-transformers
# Background task queue
celery[redis]
# Agent / LLM
langchain-google-genai
langchain-core
pytest
pytest-asyncio
